Move unit tests into `test/unit`
Change-type: patch
