+++ package.json
@@ -23,8 +23,8 @@
     "lint": "balena-lint lib test && deplint",
     "lint:fix": "balena-lint --fix lib test",
     "test": "npm run lint && npm run test:unit",
-    "test:unit": "jest lib",
-    "test:integration": "jest test",
+    "test:unit": "jest test/unit",
+    "test:integration": "jest test/integration",
     "test:compose": "docker build -t balena/jellyfish-sut:latest . && docker-compose -f docker-compose.test.yml -f docker-compose.yml up --exit-code-from=sut",
     "doc": "typedoc lib/ && touch docs/.nojekyll",
     "prepack": "npm run build",
