+++ test/unit/backend/postgres/cards.spec.ts
@@ -1,6 +1,6 @@
 import { v4 as uuid } from 'uuid';
-import * as cards from './cards';
-import { Context } from '../../context';
+import * as cards from '../../../../lib/backend/postgres/cards';
+import { Context } from '../../../../lib/context';
 
 const TEST_CONTEXT = new Context({ id: `UNIT-TEST-${uuid()}` });
